refactor(admin): extract UsageTab into its own file
